refactor: move DeleteConfirmationModal to his own module
